Enable freeze computations for rendering via GitHub Actions
This setting in the _quarto.yml is needed to allow for rendering
and publishing of the Quarto book project via GitHub Actions.
However, with this setting, execution of the R code in the project
still needs to be done locally. Maybe I will automate this at a later
stage, too.