Excel table for finding patterns among reference versions shapefiles
This commit introduces a table that supports finding patterns and
making decisions for each country on which versions of the shape-
files from EuroGeographics (v5.0/2011 or v7.0/2012) represent
their reference boundaries. This excel file compares the amounts of
LAUs with population data for each country with the amounts of LAUs
included in the two different shapefile versions.

It is constructed by code in the `prep-pop-data.R` skript and then
formatted and annotated by hand to document individual thoughts and
consideration. The line producing the Excel in the skript is
commented out to avoid that the annoteated Excel file is
overwritten by its plain version again.