Add R code to account for minor peciularities in pop data and shapefiles
The code added to `prep-pop-data.R` script with this commit, helps
to account for disparities in the two shapefile for the Budapest
LAU boundary. It also cleans out two French LAUs that do not have
unique identifier in the historic population data set, as well as
aggregates the population figures of all Budapest districts that
are initially reported in separate figures.